Only attempt to split service fields for conn logs
The conn.log is the only weird log where what should be a vector is
logged as a comma separated string.  Only attempt to split the service
field into a list if 'conn' is in the filename.  This will prevent this
from messing with the service field in the kerberos log.